Updated method config syntax to be array
Following the array shape for all configs... reduction.

Also, created a second bookshelf instance with models already registered (for native registration testing). Renamed to "service" to be more elegant with the naming.

Started the function that will load models already registered on the service, in the constructor. Thus, we can start outputting the results of this process to the console.